refactor: extract page modules (sector.py, company.py) and slim app.py
